requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.2
frozenlist==1.3.3
google-api-core==2.11.0
google-auth==2.17.3
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
googleapis-common-protos==1.59.0
grpcio==1.54.0
grpcio-status==1.54.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.24.3
openai==0.27.6
priority==2.0.0
proto-plus==1.22.2
protobuf==4.22.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dotenv==1.0.0
python-telegram-bot==20.2
pytz==2023.3
quart==0.18.4
requests==2.30.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
toml==0.10.2
tqdm==4.65.0
urllib3==1.26.6
Werkzeug==2.3.3
wsproto==1.2.0
yarl==1.9.2
zipp==3.15.0
